reference-app/frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Radim Cajzl <radim.cajzl@eon.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
requests = "^2.31.0"
gunicorn = "^20.1.0"
prometheus-flask-exporter = "^0.22.4"
opentelemetry-distro = "^0.39b0"
opentelemetry-exporter-jaeger-thrift = "^1.18.0"
opentelemetry-instrumentation-flask = "^0.39b0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
